remove 'readProperties' option, fix tests in either case
setting readProperties to FALSE and then calling getAudioProperties resulted in segmentation fault

only was an option in TagLibMPEG and there wasn't a test written for it, but the tests were messed up
in any case.

The TagLib API has a lot of optional stuff and options are nice but programming around the edgecases
only brings pain.